Fix style (extract duplicate test helper function)
